# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DrawStatus = typeof DrawStatus[keyof typeof DrawStatus];


export const DrawStatus = {
  upcoming: 'upcoming',
  active: 'active',
  completed: 'completed',
} as const;

export interface Draw {
  id: number;
  drawNumber: number;
  status: DrawStatus;
  jackpotAmount: number;
  winningTicketCode?: string | null;
  winningNumbers?: number[] | null;
  scheduledAt: string;
  drawnAt?: string | null;
  totalTicketsSold: number;
  prizePool: number;
}

export interface CreateDrawInput {
  /** @minimum 0 */
  jackpotAmount: number;
  scheduledAt: string;
}

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  available: 'available',
  sold: 'sold',
  validated: 'validated',
  claimed: 'claimed',
  expired: 'expired',
} as const;

export interface Ticket {
  id: number;
  code: string;
  status: TicketStatus;
  price: number;
  series: string;
  drawId?: number | null;
  drawNumber?: number | null;
  isWinner?: boolean;
  prizeAmount?: number | null;
  vendorId?: number | null;
  vendorName?: string | null;
  soldAt?: string | null;
  validatedAt?: string | null;
  claimedAt?: string | null;
}

export interface ValidateTicketInput {
  vendorId: number;
}

export interface ClaimPrizeInput {
  vendorId: number;
}

export interface ClaimResult {
  ticket: Ticket;
  prizeAmount: number;
  message: string;
}

export type VendorStatus = typeof VendorStatus[keyof typeof VendorStatus];


export const VendorStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Vendor {
  id: number;
  name: string;
  location: string;
  phone?: string | null;
  status: VendorStatus;
  totalSales?: number;
  totalRevenue?: number;
}

export interface CreateVendorInput {
  name: string;
  location: string;
  phone?: string;
}

export interface Stats {
  totalTicketsSold: number;
  totalRevenue: number;
  totalPrizesPaid: number;
  activeDraw?: Draw | null;
  completedDraws: number;
  activeVendors: number;
  recentWinners: number;
}

export interface Winner {
  id: number;
  ticketCode: string;
  maskedCode: string;
  prizeAmount: number;
  drawNumber: number;
  vendorName?: string | null;
  claimedAt: string;
}

export type ListDrawsParams = {
status?: ListDrawsStatus;
limit?: number;
};

export type ListDrawsStatus = typeof ListDrawsStatus[keyof typeof ListDrawsStatus];


export const ListDrawsStatus = {
  upcoming: 'upcoming',
  active: 'active',
  completed: 'completed',
} as const;

export type ListTicketsParams = {
status?: ListTicketsStatus;
limit?: number;
};

export type ListTicketsStatus = typeof ListTicketsStatus[keyof typeof ListTicketsStatus];


export const ListTicketsStatus = {
  available: 'available',
  sold: 'sold',
  validated: 'validated',
  claimed: 'claimed',
  expired: 'expired',
} as const;

export type ListWinnersParams = {
limit?: number;
};