# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  description: Halgo Cash lottery API
  version: "0.1.0"

servers:
  - url: /api

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /draws:
    get:
      operationId: listDraws
      tags: [draws]
      summary: List all draws
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of draws
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draw"
    post:
      operationId: createDraw
      tags: [draws]
      summary: Create a new draw (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrawInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
        "400":
          description: Invalid input

  /draws/latest:
    get:
      operationId: getLatestDraw
      tags: [draws]
      summary: Get the most recent completed draw
      responses:
        "200":
          description: Latest draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
        "404":
          description: No completed draw found

  /draws/{id}:
    get:
      operationId: getDraw
      tags: [draws]
      summary: Get a draw by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
        "404":
          description: Draw not found

  /draws/{id}/run:
    post:
      operationId: runDraw
      tags: [draws]
      summary: Execute a draw — pick winner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
        "400":
          description: Draw cannot be run
        "404":
          description: Draw not found

  /tickets/{code}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket status by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found

  /tickets/{code}/validate:
    post:
      operationId: validateTicket
      tags: [tickets]
      summary: Validate (sell) a ticket at POS — vendor action
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateTicketInput"
      responses:
        "200":
          description: Ticket validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Ticket already validated or invalid
        "404":
          description: Ticket not found

  /tickets/{code}/claim:
    post:
      operationId: claimPrize
      tags: [tickets]
      summary: Claim a prize for a winning ticket — vendor action
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimPrizeInput"
      responses:
        "200":
          description: Prize claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
        "400":
          description: Ticket not a winner or already claimed
        "404":
          description: Ticket not found

  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, sold, validated, claimed, expired]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"

  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get global platform statistics
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /winners:
    get:
      operationId: listWinners
      tags: [winners]
      summary: Get recent prize winners (public display)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent winners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Winner"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Draw:
      type: object
      required: [id, drawNumber, status, jackpotAmount, scheduledAt, totalTicketsSold, prizePool]
      properties:
        id:
          type: integer
        drawNumber:
          type: integer
        status:
          type: string
          enum: [upcoming, active, completed]
        jackpotAmount:
          type: number
        winningTicketCode:
          type: string
          nullable: true
        winningNumbers:
          type: array
          items:
            type: integer
          nullable: true
        scheduledAt:
          type: string
          format: date-time
        drawnAt:
          type: string
          format: date-time
          nullable: true
        totalTicketsSold:
          type: integer
        prizePool:
          type: number

    CreateDrawInput:
      type: object
      required: [jackpotAmount, scheduledAt]
      properties:
        jackpotAmount:
          type: number
          minimum: 0
        scheduledAt:
          type: string
          format: date-time

    Ticket:
      type: object
      required: [id, code, status, price, series]
      properties:
        id:
          type: integer
        code:
          type: string
        status:
          type: string
          enum: [available, sold, validated, claimed, expired]
        price:
          type: number
        series:
          type: string
        drawId:
          type: integer
          nullable: true
        drawNumber:
          type: integer
          nullable: true
        isWinner:
          type: boolean
        prizeAmount:
          type: number
          nullable: true
        vendorId:
          type: integer
          nullable: true
        vendorName:
          type: string
          nullable: true
        soldAt:
          type: string
          format: date-time
          nullable: true
        validatedAt:
          type: string
          format: date-time
          nullable: true
        claimedAt:
          type: string
          format: date-time
          nullable: true

    ValidateTicketInput:
      type: object
      required: [vendorId]
      properties:
        vendorId:
          type: integer

    ClaimPrizeInput:
      type: object
      required: [vendorId]
      properties:
        vendorId:
          type: integer

    ClaimResult:
      type: object
      required: [ticket, prizeAmount, message]
      properties:
        ticket:
          $ref: "#/components/schemas/Ticket"
        prizeAmount:
          type: number
        message:
          type: string

    Vendor:
      type: object
      required: [id, name, location, status]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        totalSales:
          type: integer
        totalRevenue:
          type: number

    CreateVendorInput:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          type: string
        phone:
          type: string

    Stats:
      type: object
      required: [totalTicketsSold, totalRevenue, totalPrizesPaid, completedDraws, activeVendors, recentWinners]
      properties:
        totalTicketsSold:
          type: integer
        totalRevenue:
          type: number
        totalPrizesPaid:
          type: number
        activeDraw:
          $ref: "#/components/schemas/Draw"
          nullable: true
        completedDraws:
          type: integer
        activeVendors:
          type: integer
        recentWinners:
          type: integer

    Winner:
      type: object
      required: [id, ticketCode, maskedCode, prizeAmount, drawNumber, claimedAt]
      properties:
        id:
          type: integer
        ticketCode:
          type: string
        maskedCode:
          type: string
        prizeAmount:
          type: number
        drawNumber:
          type: integer
        vendorName:
          type: string
          nullable: true
        claimedAt:
          type: string
          format: date-time