# lib/api-zod/src/generated/types/claimResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */
import type { Ticket } from './ticket';

export interface ClaimResult {
  ticket: Ticket;
  prizeAmount: number;
  message: string;
}