# lib/api-zod/src/generated/types/createDrawInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDrawInput {
  /** @minimum 0 */
  jackpotAmount: number;
  scheduledAt: Date;
}