# lib/api-zod/src/generated/types/draw.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */
import type { DrawStatus } from './drawStatus';

export interface Draw {
  id: number;
  drawNumber: number;
  status: DrawStatus;
  jackpotAmount: number;
  winningTicketCode?: string | null;
  winningNumbers?: number[] | null;
  scheduledAt: Date;
  drawnAt?: Date | null;
  totalTicketsSold: number;
  prizePool: number;
}