# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */
import type { Draw } from './draw';

export interface Stats {
  totalTicketsSold: number;
  totalRevenue: number;
  totalPrizesPaid: number;
  activeDraw?: Draw | null;
  completedDraws: number;
  activeVendors: number;
  recentWinners: number;
}