# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketStatus } from './ticketStatus';

export interface Ticket {
  id: number;
  code: string;
  status: TicketStatus;
  price: number;
  series: string;
  drawId?: number | null;
  drawNumber?: number | null;
  isWinner?: boolean;
  prizeAmount?: number | null;
  vendorId?: number | null;
  vendorName?: string | null;
  soldAt?: Date | null;
  validatedAt?: Date | null;
  claimedAt?: Date | null;
}