# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halgo Cash lottery API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorStatus } from './vendorStatus';

export interface Vendor {
  id: number;
  name: string;
  location: string;
  phone?: string | null;
  status: VendorStatus;
  totalSales?: number;
  totalRevenue?: number;
}